pyproject.toml
[tool.poetry]
name = "ska-tmc-integration"
version = "0.6.10"
description = "Repository for integration of TMC components"
authors = ["Team Himalaya", "Team Sahyadri"]
license = "BSD-3-Clause"

include = [
    { path = 'tests'}
]

[[tool.poetry.source]]
name = 'ska-nexus'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[tool.poetry.dependencies]
python = "3.10.11"
ska-control-model="0.3.1"
ska-tango-base="0.17.0"
numpy = "1.23.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pylint = "^2.12.2"
pytest = "^7.1"
pytest-bdd = "^5.0.0"
pytest-cov = "^2.10.1"
pylint-junit = "^0.3.2"
coverage = "^6.2"
pytest-json-report = "^1.4.1"
pytest-forked = "^1.3.0"
pytest-mock = "^3.6.1"
pytest-xdist = "^2.4.0"
pytest-repeat = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
